EditorState++
cleaned up widget object tools, a sub section of the editor.
editor state is now being modeled using an enum that more
clearly maps to the idea of 3 different states: interactive
objs, indestructible objs, and actors. rn, we just model
ints and actors, and was done using a boolean, which was
not going to scale - at all.