Indestrutibles++
complete writing routine in pool.update() to collision
check the projetile pool against the indestructiblel pool.
this finishes indestructible obj implementation. done,
tested - chain exlode is really fun to cast against inds.
the emitter stops movement and just constantly spawns
explosions in a single tile, leading to a case where the
player could get a boss against a wall and stack-up
explosions, dealing 2 damage every few frames leading
to lots of DPS, if the actor hit direction is controlled.
this would have to be between two objs that cant be
destroyed (pushable) or indestructible objs - so note:
dont make boss or hub rooms with objects that can create
traps for the bosses to be explosion-locked in. duh.