Bombos + System Maximum Tests
the current version of bombos creates explosions sequentially
left to right across the screen. this screen space is projected
down into world space (position XY), and an explosion is placed
there. for each explosion, a ground fire is created during it's
lifetime - and that ground fire creates rising smoke particles
randomly during it's lifetime - leading to a cascading state
where both the projectile and particle pools are overloaded.

this means explosions dont spawn, fires dont spawn, smoke doesn't
spawn - the pools fail gracefully, simply returning index 0
for pros and parts, but the net result is that bombos 'cuts out'
about halfway across the screen, not functioning properly.

so the fix is easy, we'll just create fire particles instead,
and explosions can burn objects in the interaction routines,
instead of relying on ground fires to handle burn for them.
this is to ensure bombos functions properly and acts as the
first (of what i hope are many) global magic spells that are
large-scale, AND functions as a max system test as well. this will
allow me to build a test level and optimize the pool interactions
even more, so the game can smoothly run large scale destruction and
interactions, which is the primary goal now that magic exists.

testing bombos using fire particles showed the pros and parts
pools hitting 90/70 during empty screen bombos. that's about 30%
of projectile pool, and 10% of the particle pool - which is a
minor upgrade from the semi-functioning version. :)