Room Wind - Part 3
we need to have more than one type of wind obj, or we need to be able
to control the wind push/speed travel values. we can create a gust with
parameters to control it's speed, and push amount, and this seems like
the best solution to keep wind as simple as possible while still allowing
for a greater degree of control.

added push amount to wind. based on assessment, wind type enum is no
longer needed. wind will always have the same sprite, it's only the
initial push that changes wind's character, plus the frequency by which
it is spawned.

consolidated push() methods in wind functions into the update(),
instead of passing the wind obj as a ref, which is much slower.
based on the wind ref's movement speed, we push the resulting
actors / objects / projectiles. the net result is hopefully faster
moving wind has more pushing power, while slower wind has less
pushing power - as one would expect.

upgraded wind spawn() with additional value to control the per frame
speed applied to the wind. this allows for even more control over the
speed of the wind once it's pushed. the level can now make 3 distinct
types of wind: gentle breezes, medium gusts, and very strong gale like
winds, all independent of frequency, entirely based on a single byte,
which is using 4 values to model these states:
0 - no wind spawns, 1 - gentle, 2 - gust, 3 - gale.

wind is working really well now.
house test level is setup to have wind moving right, with 2% spawn rate,
at a 1 intensity, or gentle breezes.