Editor Expansion to v1.0 scale
this simply means i'm defining and allocating widgets for each of the
7 islands: forest, mtn, swamp, lava, cloud, thieves den, shadow. in
addition, other widgets for both int and ind objs will remain, including
the world obj, water obj, house obj, npc obj, dev obj, and dungeon obj
widgets. so, we're going from 1 set of widgets for ints and inds to two
sets, so we can just drop in the objs when they have been defined and
test them in the editor.

moved int objs around in prep.
created two sets of 7 widgets, for interactive objs and indestructible
objs, themed to islands and general world sets. iterating these sets
was tested and refined until working properly. there is ALOT of space
for objects in the editor now.