Managing Codebase Scale
we need to be able to link and unlink all roomdata from project
without breaking the codebase. this allows us to test and dev
faster, without A TON of roomData. how to accomplish? have two
sets of roomData, one empty and one filled. class names are the
same, filenames are different. link to the data you want to for
testing and development, then link to level data for releases.
this doesn't affect the xml to cs pipeline, because we're going
to manually create these empty files, with _empty suffixes. this
prevents visual studio from slowing down when it has to process
15,000 lines of level data, plus this reduces the number of errors
displayed in the error list, allowing me to keep abusing it for
the rest of development. being smart is rewarding and fun.

i fully expect dungeon room data and field room data to reach
over a million lines alone, when game reaches version 1.0, so
being able to still build the project quickly will become vital.
empty levels still allow for testing of all game systems, via
the editor. think of this as the 'light-weight' vresion of the
codebase. we link the roomData then test and release.

lets actually do this.

created empty version of all level data files as .cs files in
the main gameclasses folder. migrated directx to use empty
level data files. building fine. sitting at 16.6K. now, the
uwp project MIGHT ACTUALLY BE USEFUL because i'm going to link
it to the actual level data, then run a code analysis on it
to determine it's size. i will continue to do this as development
progresses, so i can chart leveldata growth and observe any
useful events that might occur or influence the codebase design.
because - the codebase is about to get really big for the UWP
project, and i will finally get to experience visual studio
attempting to manage large amounts of code. the last time, vs
took me by suprise - and i had to change my codebase to work
around it's limitations. this time, i fully expect visual
studio to fail in some way, and i'm ready for it.

updated notes in uwp project folder to explain what the uwp
projects role is in the development process - as a release
testing setup only. uwp's loc is 20.3K.

comparison:
DirectX - 16.6K LoC
UWP - 20.3K LoC
(about 4K level data)