Dungeons++
we need to update dungeon / room building
to reference proper room data based on roomID.
this is where-ever forest dungeon data is.
this has been done, along with the supporting
changes. recompiled all room data.

all dungeons are ready to be tested.
we need dungeon entrances to properly
test though - and we need to change the
dungeon sheets so we have a visual diff.

we can use the black shadow animFrame for
all dungeon entrances for testing purposes.
dun ents should be located at the shadow king
level for easy testing rn.

all forest related big teeth sprites and
ind objs have been removed.coliseum bricks
have been removed as well.

we have dungeon entrances for forest, mountain,
and swamp. we need cloud, lava, thieves den, and
shadow. these are all the big shadows rn for dev.
we have these. they are now in editor, however we
need to setup their hitboxes and hide the other
objs that they cover. done.

then we need to add all the dungeon entrances
to the shadowking level, and build little shrines
around them so we know which ones are which.
save shadowking level. done.

then we need to recompile room data and test
that we can enter and explore each dungeon type.
we need to have hero interaction rec with all
dungeon entrances and exits. done. tested and
working. very good.