Level Presentation Ideas
themeing levels using objects + massive sprites.
these are ints and inds, plus a sprite component.
this sprite component sits along the TOP of the level,
sitting above the middle of the level, so it presents
nicely to the player. it is the width of the level, plus
16*2 pixels. since this will become square in the process,
we should go ahead and make this square in photoshop.
this allows us to also have custom sprites that look
like objects south of the entrance. at this point, we're
basically painting the entrance and it's custom surrounding
objects, then layering on inds to complete the hitboxes,
then adding more inds and int objects that the player can
interact with to complete the level.

where do we store the massive sprite? level screen? pool?
what is this big sprite called?

PresentationSprite

this is the big sprite's name, as it also describes it's
purpose. it's meant to present the level, whatever it may
be, in a grand and breath-taking 1995 nintendo style way.
but just a little bit further than that, because they were
working with hardware limitations and time constraints,
whereas i don't have either of those.

screenshotted field level width to a photoshop file,
cropped and setup file for creating presentation sprite.
this is a 1300x400 sprite. this allows for about half of the
level to be custom hand-painted pixel art, with additional
objects layered on top (with a system designed to handle
an entire levels worth of interactive objects). so, things
are looking good....

where does this sprite go? this is a single sprite
component, always placed at the same position on a level.
i think it should be part of level, that way we dont
expand room data. the sprite can change based on the
field id. we would never draw the sprite or use it in
a dungeon room, for now (unless we resize boss rooms
to be very wide and then use the sprite to add even more
flavor to the battle). which, we could do that. sure.
note of this idea added to ideas.