Presentation Sprite v1
handled by level, this is a 1300x400 texture that sits at
the same location in every room. for dungeeons, this
presentation sprite is only drawn in the hub and boss rooms.
for field levels, this presentation sprite is drawn for every
level. it's not always needed, but it can be used.

we will use a reference that points to a texture instance,
based on room.id, when the room is built. we need 1 sprite
component in level screen that draws before the ind objs.
when the level builds, we will swap it's texture based on
the roomID we're using the build the room. this works for
fields and dungeon rooms, like hub and boss rooms.

so, based on the testing for presentation sprite, it
kinda makes more sense to call it a background sprite,
because we layer everything on top of it. this sprite
cannot contain objects that have depth, tho, so we can
do far north walls, and then middle area ground effects,
or objects that have height that extend to the wall,
or objs that have height that are covered above somehow.
we can also have objs that sit over the bkgsprite that
cover link, so it appears he is 'inside' the bkg sprite,
but that is very specific to the level. overall testing
is good.

we dev tested on a mountin wall, and this will work.
so let's make the coliseum entrance as an in-game test.

made dev test coliseum entrance based on current coliseum
entrance shrine. exported to png, added to dx project for
testing in the shadow coliseum level.

tested and working for the coliseum level. the bkg
sprite is basic but shows the new direction important
levels need to take.

bkg sprite has been moved to levelset class for easier
organization. working as expected.