BkgSprite++
we need a grid on the bkg sprite master file.
this needs to line up to the level EXACTLY.
so we need to print screen and position until
we have the grid lined up perfectly. we also need
to place some objs in the scene and print screen
them into the bkg sprite master file, so we can
build the grid based on the game world grid,
making sure they are aligned, because then we
can remove the stair refs and merge the bkg sprite
with the int objs we have already to make the coliseum
actually functioning and working visually too.

done, working and looking great.

fixed a debug path bug with uwp too.